projetos/data_science/nlp/sentiment_analysis_dataset/requirements.txt
pandas
joblib
scikit-learn
nltk
seaborn
matplotlib
streamlit
openai-whisper
Unidecode

